requirements_dev.txt
pip==9.0.1
bumpversion==0.5.3
wheel==0.30.0
watchdog==0.8.3
flake8==3.4.1
tox==2.8.2
coverage==4.4.1
Sphinx==1.6.3
cryptography==2.0.3
PyYAML==3.12
pytest==3.2.2
pytest-runner==2.12.1
vcrpy==1.11.1
